+++ Changelog.md
@@ -1,6 +1,10 @@
 # Plugin Machine CLI Changelog
 
-# 0.4.1
+## 0.5.1
+
+- Load rules based on local path
+
+## 0.4.1
 
 - Log what needs to get added to main plugin file.
 - More descriptive errors in CLI output when code API returns error